broadcast changes when committing from pg endpoint
