+++ crates/corro-agent/src/api/public/mod.rs
@@ -1,8 +1,4 @@
-use std::{
-    iter::Peekable,
-    ops::RangeInclusive,
-    time::{Duration, Instant},
-};
+use std::time::{Duration, Instant};
 
 use axum::{response::IntoResponse, Extension};
 use bytes::{BufMut, BytesMut};
@@ -11,7 +7,7 @@ use corro_types::{
     agent::{Agent, ChangeError, KnownDbVersion},
     api::{row_to_change, ExecResponse, ExecResult, QueryEvent, Statement},
     broadcast::{ChangeV1, Changeset, Timestamp},
-    change::SqliteValue,
+    change::{ChunkedChanges, SqliteValue, MAX_CHANGES_BYTE_SIZE},
     schema::{apply_schema, parse_sql},
     sqlite::SqlitePoolError,
 };
@@ -29,124 +25,11 @@ use tokio::{
 };
 use tracing::{debug, error, info, trace};
 
-use corro_types::{
-    broadcast::{BroadcastInput, BroadcastV1},
-    change::Change,
-};
+use corro_types::broadcast::{BroadcastInput, BroadcastV1};
 
 use crate::agent::process_subs;
 
 pub mod pubsub;
-
-pub struct ChunkedChanges<I: Iterator> {
-    iter: Peekable<I>,
-    changes: Vec<Change>,
-    last_pushed_seq: i64,
-    last_start_seq: i64,
-    last_seq: i64,
-    max_buf_size: usize,
-    buffered_size: usize,
-    done: bool,
-}
-
-impl<I> ChunkedChanges<I>
-where
-    I: Iterator,
-{
-    pub fn new(iter: I, start_seq: i64, last_seq: i64, max_buf_size: usize) -> Self {
-        Self {
-            iter: iter.peekable(),
-            changes: vec![],
-            last_pushed_seq: 0,
-            last_start_seq: start_seq,
-            last_seq,
-            max_buf_size,
-            buffered_size: 0,
-            done: false,
-        }
-    }
-
-    pub fn max_buf_size(&self) -> usize {
-        self.max_buf_size
-    }
-
-    pub fn set_max_buf_size(&mut self, size: usize) {
-        self.max_buf_size = size;
-    }
-}
-
-impl<I> Iterator for ChunkedChanges<I>
-where
-    I: Iterator<Item = rusqlite::Result<Change>>,
-{
-    type Item = Result<(Vec<Change>, RangeInclusive<i64>), rusqlite::Error>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        // previously marked as done because the Rows iterator returned None
-        if self.done {
-            return None;
-        }
-
-        debug_assert!(self.changes.is_empty());
-
-        // reset the buffered size
-        self.buffered_size = 0;
-
-        loop {
-            trace!("chunking through the rows iterator");
-            match self.iter.next() {
-                Some(Ok(change)) => {
-                    trace!("got change: {change:?}");
-
-                    self.last_pushed_seq = change.seq;
-
-                    self.buffered_size += change.estimated_byte_size();
-
-                    self.changes.push(change);
-
-                    if self.last_pushed_seq == self.last_seq {
-                        // this was the last seq! break early
-                        break;
-                    }
-
-                    if self.buffered_size >= self.max_buf_size {
-                        // chunking it up
-                        let start_seq = self.last_start_seq;
-
-                        if self.iter.peek().is_none() {
-                            // no more rows, break early
-                            break;
-                        }
-
-                        // prepare for next round! we're not done...
-                        self.last_start_seq = self.last_pushed_seq + 1;
-
-                        return Some(Ok((
-                            self.changes.drain(..).collect(),
-                            start_seq..=self.last_pushed_seq,
-                        )));
-                    }
-                }
-                None => {
-                    // probably not going to happen since we peek at the next and end early
-                    // break out of the loop, don't return, there might be buffered changes
-                    break;
-                }
-                Some(Err(e)) => return Some(Err(e)),
-            }
-        }
-
-        self.done = true;
-
-        // return buffered changes
-        Some(Ok((
-            self.changes.clone(),                // no need to drain here like before
-            self.last_start_seq..=self.last_seq, // even if empty, this is all we have still applied
-        )))
-    }
-}
-
-const MAX_CHANGES_BYTE_SIZE: usize = 8 * 1024;
 
 pub async fn make_broadcastable_changes<F, T>(
     agent: &Agent,
@@ -1147,124 +1030,5 @@ mod tests {
         assert_eq!(count, 4);
 
         Ok(())
-    }
-
-    #[test]
-    fn test_change_chunker() {
-        // empty interator
-        let mut chunker = ChunkedChanges::new(vec![].into_iter(), 0, 100, 50);
-
-        assert_eq!(chunker.next(), Some(Ok((vec![], 0..=100))));
-        assert_eq!(chunker.next(), None);
-
-        let changes: Vec<Change> = (0..100)
-            .map(|seq| Change {
-                seq,
-                ..Default::default()
-            })
-            .collect();
-
-        // 2 iterations
-        let mut chunker = ChunkedChanges::new(
-            vec![
-                Ok(changes[0].clone()),
-                Ok(changes[1].clone()),
-                Ok(changes[2].clone()),
-            ]
-            .into_iter(),
-            0,
-            100,
-            changes[0].estimated_byte_size() + changes[1].estimated_byte_size(),
-        );
-
-        assert_eq!(
-            chunker.next(),
-            Some(Ok((vec![changes[0].clone(), changes[1].clone()], 0..=1)))
-        );
-        assert_eq!(
-            chunker.next(),
-            Some(Ok((vec![changes[2].clone()], 2..=100)))
-        );
-        assert_eq!(chunker.next(), None);
-
-        let mut chunker = ChunkedChanges::new(
-            vec![Ok(changes[0].clone()), Ok(changes[1].clone())].into_iter(),
-            0,
-            0,
-            changes[0].estimated_byte_size(),
-        );
-
-        assert_eq!(chunker.next(), Some(Ok((vec![changes[0].clone()], 0..=0))));
-        assert_eq!(chunker.next(), None);
-
-        // gaps
-        let mut chunker = ChunkedChanges::new(
-            vec![Ok(changes[0].clone()), Ok(changes[2].clone())].into_iter(),
-            0,
-            100,
-            changes[0].estimated_byte_size() + changes[2].estimated_byte_size(),
-        );
-
-        assert_eq!(
-            chunker.next(),
-            Some(Ok((vec![changes[0].clone(), changes[2].clone()], 0..=100)))
-        );
-
-        assert_eq!(chunker.next(), None);
-
-        // gaps
-        let mut chunker = ChunkedChanges::new(
-            vec![
-                Ok(changes[2].clone()),
-                Ok(changes[4].clone()),
-                Ok(changes[7].clone()),
-                Ok(changes[8].clone()),
-            ]
-            .into_iter(),
-            0,
-            100,
-            100000, // just send them all!
-        );
-
-        assert_eq!(
-            chunker.next(),
-            Some(Ok((
-                vec![
-                    changes[2].clone(),
-                    changes[4].clone(),
-                    changes[7].clone(),
-                    changes[8].clone()
-                ],
-                0..=100
-            )))
-        );
-
-        assert_eq!(chunker.next(), None);
-
-        // gaps
-        let mut chunker = ChunkedChanges::new(
-            vec![
-                Ok(changes[2].clone()),
-                Ok(changes[4].clone()),
-                Ok(changes[7].clone()),
-                Ok(changes[8].clone()),
-            ]
-            .into_iter(),
-            0,
-            10,
-            changes[2].estimated_byte_size() + changes[4].estimated_byte_size(),
-        );
-
-        assert_eq!(
-            chunker.next(),
-            Some(Ok((vec![changes[2].clone(), changes[4].clone(),], 0..=4)))
-        );
-
-        assert_eq!(
-            chunker.next(),
-            Some(Ok((vec![changes[7].clone(), changes[8].clone(),], 5..=10)))
-        );
-
-        assert_eq!(chunker.next(), None);
     }
 }
