+++ crates/corro-pg/Cargo.toml
@@ -10,6 +10,8 @@ corro-types = { path = "../corro-types" }
 fallible-iterator = { workspace = true }
 futures = { workspace = true }
 hex = { workspace = true }
+itertools = { workspace = true }
+metrics = { workspace = true }
 pgwire = { version = "0.16.1" }
 postgres-types = { version = "0.2", features = ["with-time-0_3"] }
 rusqlite = { workspace = true }
